Rename `IcyMetadataStreamParser` to `IcyMetadataBuffer` and simplify metadata handling logic by removing unused properties and methods.
